Add: moved UpdateBalance function to AccountContext
