+++ src/components/CreateCampaign/CampaignSummary.tsx
@@ -11,9 +11,7 @@ import { ConfirmModal, SuccessModal } from 'components/common/Modals'
 import AttentionIcon from 'resources/icons/Attention'
 import useCampaignsData from 'hooks/useCampaignsData'
 import useCustomNotifications from 'hooks/useCustomNotifications'
-import { Account } from 'types'
 import useAccount from 'hooks/useAccount'
-import { useAdExApi } from 'hooks/useAdexServices'
 
 const useStyles = createStyles((theme) => ({
   bg: {
@@ -57,8 +55,7 @@ const useStyles = createStyles((theme) => ({
 const CampaignSummary = () => {
   const { classes, cx } = useStyles()
   const [opened, { open, close }] = useDisclosure(false)
-  const { adexAccount, updateAdexAccount } = useAccount()
-  const { adexServicesRequest } = useAdExApi()
+  const { updateBalance } = useAccount()
   const {
     campaign: { step, adUnits },
     updateCampaign,
@@ -93,27 +90,6 @@ const CampaignSummary = () => {
 
   const isTheLastStep = useMemo(() => step === CREATE_CAMPAIGN_STEPS - 1, [step])
   const isFirstStep = useMemo(() => step === 0, [step])
-  const updateBalance = useCallback(async () => {
-    try {
-      const getBalance = await adexServicesRequest<Account>('backend', {
-        route: '/dsp/accounts/my-account',
-        method: 'GET'
-      })
-
-      if (getBalance) {
-        updateAdexAccount({ ...adexAccount, ...getBalance })
-      } else {
-        showNotification(
-          'error',
-          'Updating account balance failed',
-          'Updating account balance failed'
-        )
-      }
-    } catch (err: any) {
-      console.error('Updating account balance failed:', err)
-      showNotification('error', err, 'Updating account balance failed')
-    }
-  }, [adexAccount, adexServicesRequest, showNotification, updateAdexAccount])
 
   const launchCampaign = useCallback(async () => {
     try {
